corrected dine in page, added ternary operators for weekday and lunch items
