+++ client/src/components/Counter.js
@@ -52,7 +52,7 @@ function Counter() {
         }
     }, [count])
 
-    if (moment().isAfter(localStorage.getItem("endtime")) || difference > 6000) {
+    if (difference > 6000) {
         return (
             <Container>
                 <div className="float-right counter">
@@ -65,7 +65,7 @@ function Counter() {
         return (
             <Container>
                 <div className="float-right counter">
-                    <div className="counter float-right" style={{ color }}>{localStorage.getItem("endtime") === null ?  " " : `${hour}:${minute}:${seconds}`} </div>
+                    <div className="counter float-right" style={{ color }}>{`${hour}:${minute}:${seconds}`} </div>
                     </div>
             </Container>
         )
