+++ client/src/components/Rules/LunchRules.js
@@ -1,0 +1,12 @@
+import React from "react";
+
+export default function LunchRule () {
+    return (
+        <>
+        <h2>Lunch (until 3PM): time limit 60 min for $24.95 (Nigiri NOT included) </h2>
+        <h3>for Kid's 2 to 7yrs: $10.95</h3>
+        <hr/>
+        </>
+    )
+}
+
